Added second round of file structure for pages and css files. Added SignUp form and created Date Picker.
