+++ propupapp/src/components/pages/Book/Book.css
@@ -1,0 +1,3 @@
+.Picker {
+    margin-top: 70px;
+}
