+++ propupapp/src/components/pages/Events/Events.js
@@ -1,0 +1,12 @@
+import React from "react";
+import './Events.css';
+
+const Events = () => {
+  return(
+    <div className="App">
+      <h1>Event Page API</h1>
+    </div>  
+    )
+};
+
+export default Events;
