+++ propupapp/src/components/pages/Owner/Owner.js
@@ -1,8 +1,9 @@
 import React from "react";
+import './Owner.css';
 
 const Owner = () => {
   return(
-    <div className="App">
+    <div className="Owner">
       <h1>Content for the home page goes right here.</h1>
       <h1>Content for the home page goes right here.</h1>
       <h1>Content for the home page goes right here.</h1>
