+++ propupapp/src/components/pages/Signup/Signup.css
@@ -1,0 +1,5 @@
+.signup-container {
+    margin-top: 80px;
+    margin-bottom: 40px;
+    column-width: 800px;
+}
