+++ propupapp/src/components/routes/Router.js
@@ -1,11 +1,12 @@
 import React, { Component } from "react";
 import { Route, Switch } from "react-router-dom";
 import Home from "../pages/Home/Home";
-import Signup from "../pages/Signup";
-import Login from "../pages/Login";
-import Owner from "../pages/Owner";
-import Renter from "../pages/Renter";
-import Book from "../pages/Book";
+import Signup from "../pages/Signup/Signup";
+import Login from "../pages/Login/Login";
+import Owner from "../pages/Owner/Owner";
+import Renter from "../pages/Renter/Renter";
+import Book from "../pages/Book/Book";
+import Events from "../pages/Events/Events";
 
 class Router extends Component {
   render() {
@@ -17,6 +18,7 @@ class Router extends Component {
         <Route component={Owner} exact path="/Owner" />
         <Route component={Renter} exact path="/Renter" />
         <Route component={Book} exact path="/Book" />
+        <Route component={Events} exact path="/Events" />
       </Switch>
     );
   }
